[util] Resolve type visibility issues.
If IapUtil is templated with <R: IapRepository>, clients can't
instantiate it because IapRepositoryImpl and the DataSourcImpl types are
private. Instead, simply fix the repository type to IapRepositoryImpl
on the util. This means we won't easily be able to test the util by
swapping out the backend, but since it's just a wrapper for the
repository, that seems fine.